@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
@@ -44,6 +44,7 @@ dependencies:
   flutter_polyline_points: ^1.0.0
   geolocator: ^7.3.1
 #  geocoder: ^0.2.1
+  geocoding: ^2.0.0
   google_maps_flutter: ^2.0.6
   google_sign_in: ^5.0.5
   google_directions_api: ^0.9.0
