@@ -14,8 +14,8 @@ import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 
 
 class RequestDeliveryPage extends StatefulWidget {
-  final DeliveryRequest request;
-  @required final UjuziUser requestingUser;
+  final DeliveryRequest? request;
+  @required final UjuziUser? requestingUser;
 
   RequestDeliveryPage({this.request, this.requestingUser});
 
@@ -25,14 +25,14 @@ class RequestDeliveryPage extends StatefulWidget {
 }
 
 class _RequestDeliveryPageState extends State<RequestDeliveryPage> {
-  String deliveryID;
-  UjuziUser requestingUser ; // the ujuzi user who requested the delivery //todo is this a duplicate of pickup person?
-  DateTime requestDate = DateTime.now(); // the date and time at which the delivery request was made
-  DateTime startDate; //the date and time at which the delivery commenced
-  DateTime completionDate; //the date and time at which the delivery completed
-  DeliveryStatus  status = DeliveryStatus.pending; // the status of the delivery
-  PackageType packageType = PackageType.parcel;
-  PaymentMethod paymentMethod = PaymentMethod.paymentOnPickup;
+  String? deliveryID;
+  UjuziUser? requestingUser ; // the ujuzi user who requested the delivery //todo is this a duplicate of pickup person?
+  DateTime? requestDate = DateTime.now(); // the date and time at which the delivery request was made
+  DateTime? startDate; //the date and time at which the delivery commenced
+  DateTime? completionDate; //the date and time at which the delivery completed
+  DeliveryStatus?  status = DeliveryStatus.pending; // the status of the delivery
+  PackageType? packageType = PackageType.parcel;
+  PaymentMethod? paymentMethod = PaymentMethod.paymentOnPickup;
   TextEditingController dropOffPersonName = new TextEditingController();
   TextEditingController dropOffPersonNumber = new TextEditingController();
   // DeliveryLocation dropOffLocation = new DeliveryLocation();
@@ -54,31 +54,31 @@ class _RequestDeliveryPageState extends State<RequestDeliveryPage> {
   void initState() {
     super.initState();
     setState(() {
-      requestingUser = widget.requestingUser ?? widget.request.requestingUser;
-      pickupPersonName.text = requestingUser.username;
-      pickupPersonNumber.text = requestingUser.number;
-      pickupLocationController.location = requestingUser.location;
-      pickupLocationController.location = requestingUser.location;
+      requestingUser = widget.requestingUser ?? widget.request!.requestingUser;
+      pickupPersonName.text = requestingUser!.username;
+      pickupPersonNumber.text = requestingUser!.number;
+      pickupLocationController.location = requestingUser!.location;
+      pickupLocationController.location = requestingUser!.location;
     });
 
 
      if (widget.request != null){
        setState(() {
 
-         requestDate = widget.request.requestDate;
-         startDate = widget.request.startDate;
-         completionDate = widget.request.completionDate;
-         status = widget.request.status;
-         packageType = widget.request.packageType;
-         dropOffPersonName.text = widget.request.dropOffPersonName;
-         dropOffPersonNumber.text = widget.request.dropOffPersonNumber;
-         dropOffLocationController.text = widget.request.dropOffLocation.locationName;
-         dropOffLocationController.location = widget.request.dropOffLocation;
-         pickupPersonNumber.text = widget.request.pickupPersonNumber;
-         pickupPersonName.text = widget.request.pickupPersonName;
-         pickupLocationController.location = widget.request.pickupLocation;
-         pickupLocationController.text = widget.request.pickupLocation.locationName;
-         notes.text = widget.request.notes;
+         requestDate = widget.request!.requestDate;
+         startDate = widget.request!.startDate;
+         completionDate = widget.request!.completionDate;
+         status = widget.request!.status;
+         packageType = widget.request!.packageType;
+         dropOffPersonName.text = widget.request!.dropOffPersonName!;
+         dropOffPersonNumber.text = widget.request!.dropOffPersonNumber!;
+         dropOffLocationController.text = widget.request!.dropOffLocation!.locationName!;
+         dropOffLocationController.location = widget.request!.dropOffLocation!;
+         pickupPersonNumber.text = widget.request!.pickupPersonNumber!;
+         pickupPersonName.text = widget.request!.pickupPersonName!;
+         pickupLocationController.location = widget.request!.pickupLocation!;
+         pickupLocationController.text = widget.request!.pickupLocation!.locationName!;
+         notes.text = widget.request!.notes!;
 
 
        });
@@ -129,12 +129,12 @@ class _RequestDeliveryPageState extends State<RequestDeliveryPage> {
 
                     header: Padding(
                       padding: const EdgeInsets.all(8.0),
-                      child: Text(AppLocalizations.of(context).pickup_person_details),
+                      child: Text(AppLocalizations.of(context)!.pickup_person_details),
                     ),
                     collapsed: Container(
                       padding: EdgeInsets.all(8),
                       width: size.width,
-                        child: Text("${requestingUser.username}, ${requestingUser.number}, ${pickupLocationController.text}"),
+                        child: Text("${requestingUser!.username}, ${requestingUser!.number}, ${pickupLocationController.text}"),
                       color: Colors.white54,
                     ),
 
@@ -149,7 +149,7 @@ class _RequestDeliveryPageState extends State<RequestDeliveryPage> {
 
 
                           CustomTextField(
-                            label: AppLocalizations.of(context).pickup_person_name,
+                            label: AppLocalizations.of(context)!.pickup_person_name,
                             color: Colors.black,
                             controller: pickupPersonName,
                             onChanged: (value){
@@ -163,12 +163,12 @@ class _RequestDeliveryPageState extends State<RequestDeliveryPage> {
                           InternationalPhoneNumberInput(
                               countries: ["CD"],
                               textFieldController: pickupPersonNumber,
-                              hintText: AppLocalizations.of(context).pickup_person_number,
+                              hintText: AppLocalizations.of(context)!.pickup_person_number,
                               validator: (value){
                                 if (value == null || value.isEmpty)
-                                  return AppLocalizations.of(context).required_field;
+                                  return AppLocalizations.of(context)!.required_field;
                                 else if (value.length <= 9)
-                                  return AppLocalizations.of(context).valid_phone_number;
+                                  return AppLocalizations.of(context)!.valid_phone_number;
                                 return "";
                               },
                               selectorConfig: SelectorConfig(
@@ -183,7 +183,7 @@ class _RequestDeliveryPageState extends State<RequestDeliveryPage> {
 
 
                           LocationTextField(
-                            label: AppLocalizations.of(context).pickup_person_location,
+                            label: AppLocalizations.of(context)!.pickup_person_location,
                             color: Colors.black,
                             focusNode: pickupLocationNode,
                             labelColor: Colors.grey,
@@ -193,7 +193,7 @@ class _RequestDeliveryPageState extends State<RequestDeliveryPage> {
                               determinePosition().then((value) {
                                 setState(() {
                                   pickupLocationController.location = value;
-                                  pickupLocationController.text = value.locationName;
+                                  pickupLocationController.text = value.locationName!;
                                 });
                               });
                               },
@@ -207,7 +207,7 @@ class _RequestDeliveryPageState extends State<RequestDeliveryPage> {
 
 
                 LocationTextField(
-                  label: AppLocalizations.of(context).where_to,
+                  label: AppLocalizations.of(context)!.where_to,
                   color: Colors.black,
                   focusNode: dropoffLocationNode,
                   labelColor: Colors.grey,
@@ -219,7 +219,7 @@ class _RequestDeliveryPageState extends State<RequestDeliveryPage> {
                     pickLocation(context).then((value) {
                       if (value != null) {
                         setState(() {
-                          dropOffLocationController.text = value.locationName;
+                          dropOffLocationController.text = value.locationName!;
                         });
                       }
                     });
@@ -235,7 +235,7 @@ class _RequestDeliveryPageState extends State<RequestDeliveryPage> {
                   child: Padding(
                     padding: const EdgeInsets.only(top: 40.0, left: 32.0),
                     child: Text(
-                      AppLocalizations.of(context).pickup_details,
+                      AppLocalizations.of(context)!.pickup_details,
                       style: TextStyle(
                         fontWeight: FontWeight.bold,
                         fontSize: 16.0
@@ -246,7 +246,7 @@ class _RequestDeliveryPageState extends State<RequestDeliveryPage> {
 
                 CustomTextField(
                   controller: dropOffPersonName,
-                  label: AppLocalizations.of(context).dropoff_person_name,
+                  label: AppLocalizations.of(context)!.dropoff_person_name,
                   color: Colors.black,
                   onChanged: (value){
                   },
@@ -263,12 +263,12 @@ class _RequestDeliveryPageState extends State<RequestDeliveryPage> {
                   child: InternationalPhoneNumberInput(
                       countries: ["CD"],
                     textFieldController: dropOffPersonNumber,
-                    hintText: AppLocalizations.of(context).dropoff_person_number,
+                    hintText: AppLocalizations.of(context)!.dropoff_person_number,
                       validator: (value){
                         if (value == null || value.isEmpty)
-                          return AppLocalizations.of(context).required_field;
+                          return AppLocalizations.of(context)!.required_field;
                         else if (value.length <= 9)
-                          return AppLocalizations.of(context).valid_phone_number;
+                          return AppLocalizations.of(context)!.valid_phone_number;
                         return "";
                       },
                       selectorConfig: SelectorConfig(
@@ -287,7 +287,7 @@ class _RequestDeliveryPageState extends State<RequestDeliveryPage> {
 
 
                 CustomTextField(
-                  label: AppLocalizations.of(context).notes,
+                  label: AppLocalizations.of(context)!.notes,
                   color: Colors.black,
                   labelColor: Colors.grey,
                   widthFactor: 0.85,
@@ -304,24 +304,24 @@ class _RequestDeliveryPageState extends State<RequestDeliveryPage> {
                   children: [
                     Padding(
                       padding: const EdgeInsets.only(left: 36.0),
-                      child: Text("${AppLocalizations.of(context).package_type}: "),
+                      child: Text("${AppLocalizations.of(context)!.package_type}: "),
                     ),
 
                     Padding(
                       padding: const EdgeInsets.only(left: 24.0),
                       child: DropdownButton(
-                        onChanged: (value){
+                        onChanged: (dynamic value){
                           setState(() {
                             packageType = value;
 
                           });
                         },
-                        hint: Text(AppLocalizations.of(context).package_type),
+                        hint: Text(AppLocalizations.of(context)!.package_type),
                         value: packageType,
                         items: [
-                          DropdownMenuItem(value: PackageType.letter, child: Text(AppLocalizations.of(context).letter)),
-                          DropdownMenuItem(value: PackageType.parcel, child: Text(AppLocalizations.of(context).parcel)),
-                          DropdownMenuItem(value: PackageType.large, child: Text(AppLocalizations.of(context).large)),
+                          DropdownMenuItem(value: PackageType.letter, child: Text(AppLocalizations.of(context)!.letter)),
+                          DropdownMenuItem(value: PackageType.parcel, child: Text(AppLocalizations.of(context)!.parcel)),
+                          DropdownMenuItem(value: PackageType.large, child: Text(AppLocalizations.of(context)!.large)),
                         ],
                       ),
                     ),
@@ -334,22 +334,22 @@ class _RequestDeliveryPageState extends State<RequestDeliveryPage> {
                   child: Row(
                     mainAxisAlignment: MainAxisAlignment.start,
                     children: [
-                      Text("${AppLocalizations.of(context).payment_method}: "),
+                      Text("${AppLocalizations.of(context)!.payment_method}: "),
 
                       Padding(
                         padding: const EdgeInsets.only(left: 12.0),
                         child: DropdownButton(
-                          onChanged: (value){
+                          onChanged: (dynamic value){
                             setState(() {
                               paymentMethod = value;
                             });
                           },
-                          hint: Text(AppLocalizations.of(context).payment_method),
+                          hint: Text(AppLocalizations.of(context)!.payment_method),
                           value: paymentMethod,
                           items: [
-                            DropdownMenuItem(value: PaymentMethod.paymentOnDelivery, child: Text(AppLocalizations.of(context).payment_on_delivery)),
-                            DropdownMenuItem(value: PaymentMethod.paymentOnPickup, child: Text(AppLocalizations.of(context).payment_on_pickup)),
-                            DropdownMenuItem(value: PaymentMethod.creditCard, child: Text(AppLocalizations.of(context).payment_via_credit_card)),
+                            DropdownMenuItem(value: PaymentMethod.paymentOnDelivery, child: Text(AppLocalizations.of(context)!.payment_on_delivery)),
+                            DropdownMenuItem(value: PaymentMethod.paymentOnPickup, child: Text(AppLocalizations.of(context)!.payment_on_pickup)),
+                            DropdownMenuItem(value: PaymentMethod.creditCard, child: Text(AppLocalizations.of(context)!.payment_via_credit_card)),
                           ],
                         ),
                       ),
@@ -358,7 +358,7 @@ class _RequestDeliveryPageState extends State<RequestDeliveryPage> {
                 ),
 
                 CustomRoundedButton(
-                  text: AppLocalizations.of(context).request_delivery.toUpperCase(),
+                  text: AppLocalizations.of(context)!.request_delivery.toUpperCase(),
                   onPressed: (){
                     DeliveryRequest request = DeliveryRequest(
                         requestingUser: requestingUser,
@@ -380,7 +380,7 @@ class _RequestDeliveryPageState extends State<RequestDeliveryPage> {
 
 
                     if (widget.request != null)//old delivery
-                      request.setReference(widget.request.reference);
+                      request.setReference(widget.request!.reference);
                     showModalBottomSheet(
                       enableDrag: false,
                       context: context,
