@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 
 
 class CustomTextButton extends StatelessWidget {
-  final String foreText;
-  @required final  String actionText;
+  final String? foreText;
+  @required final  String? actionText;
   @required final onPressed;
 
   CustomTextButton({
