@@ -2,14 +2,14 @@ import 'package:ujuzi_xpress/utils/models/UjuziUser.dart';
 
 
 abstract class LoginState {
-  UjuziUser user;
-  String message;
+  UjuziUser? user;
+  String? message;
 
   LoginState({this.user, this.message});
 }
 
 class LoginAuthenticated extends LoginState {
-  UjuziUser user;
+  UjuziUser? user;
   LoginAuthenticated(this.user) : super(user: user);
 }
 
@@ -17,6 +17,6 @@ class LoginUnauthenticated extends LoginState {}
 
 
 class LoginError extends LoginState {
-  String message;
+  String? message;
   LoginError(this.message) : super(message: message);
 }
