@@ -1,9 +1,9 @@
 
 
 class Rider{
-  String _name;
-  String _phoneNumber;
-  String _profileImageUrl;
+  String? _name;
+  String? _phoneNumber;
+  String? _profileImageUrl;
 
   String get name => this._name ?? "";
   String get phoneNumber => this._phoneNumber ?? "";
