@@ -6,19 +6,18 @@ import 'package:flutter/material.dart';
 class Directions {
   final LatLngBounds bounds;
   final List<PointLatLng> polylinePoints;
-  final String totalDistance;
-  final String totalDuration;
+  final String? totalDistance;
+  final String? totalDuration;
 
   const Directions({
-    @required this.bounds,
-    @required this.polylinePoints,
-    @required this.totalDistance,
-    @required this.totalDuration,
+    required this.bounds,
+    required this.polylinePoints,
+    required this.totalDistance,
+    required this.totalDuration,
   });
 
   factory Directions.fromMap(Map<String, dynamic> map) {
     // Check if route is not available
-    if ((map['routes'] as List).isEmpty) return null;
 
     // Get route information
     final data = Map<String, dynamic>.from(map['routes'][0]);
@@ -32,8 +31,8 @@ class Directions {
     );
 
     // Distance & Duration
-    String distance = '';
-    String duration = '';
+    String? distance = '';
+    String? duration = '';
     if ((data['legs'] as List).isNotEmpty) {
       final leg = data['legs'][0];
       distance = leg['distance']['text'];
